exp-management/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.4.0
pytest-django==4.5.2
mongoengine==0.27.0
pymongo==4.3.3
dnspython==2.3.0
